2022-11: adjust for secondary puzzle
this one was tough. Numbers got huge without the `div` 3 from first part.
The trick was to find a common denominator for all denominators (divisors).

this is descibed here: https://en.wikipedia.org/wiki/Chinese_remainder_theorem
although the puzzle does not explicitly state that all denominators will
always be primes, so the solution will not scale to arbitrary inputs.